feat(avanues): Expandable command drill-down, PulseDot refinement, scroll reclassification
- Command rows now expand to show Primary phrase, Synonyms chips, Function label, and Add Alias
- PulseDot: Running uses calm glow, Error/Degraded use single ring (less distracting)
- Moved scroll commands from Accessibility to Navigation category
- Fixed Color(ULong) crash in accessibility service and cursor settings (use Color(Int) for ARGB)
- Fixed IntrinsicSize.Min crash with SubcomposeLayout in module cards
- Added registered trademark superscript in dashboard header
- Fixed module card sizing with maxLines=1 for status text

+++ apps/avanues/src/main/kotlin/com/augmentalis/voiceavanue/service/VoiceAvanueAccessibilityService.kt
@@ -247,7 +247,7 @@ class VoiceAvanueAccessibilityService : VoiceOSAccessibilityService() {
                         // Restore custom accent override from persisted value
                         val accentOverride = settings.cursorAccentOverride
                         if (accentOverride != null) {
-                            val color = androidx.compose.ui.graphics.Color(accentOverride.toULong())
+                            val color = androidx.compose.ui.graphics.Color(accentOverride.toInt())
                             AvanueModuleAccents.setOverride(
                                 "voicecursor",
                                 ModuleAccent(
